Add cleanup jobs for GCP folders, projects, and instances. Separate compute module to own python file.
