+++ cartography/data/jobs/cleanup/gcp_compute_instance_cleanup.json
@@ -1,0 +1,17 @@
+{
+  "statements": [
+    {
+      "query": "MATCH (n:GCPInstance) WHERE n.lastupdated <> {UPDATE_TAG} WITH n LIMIT {LIMIT_SIZE} DETACH DELETE (n) return COUNT(*) as TotalCompleted",
+      "iterative": true,
+      "iterationsize": 100,
+      "__comment__": "Delete GCP Instances that no longer exist and detach them from all previously connected nodes."
+    },
+    {
+      "query": "MATCH (:GCPInstance)<-[r:RESOURCE]-(:GCPProject) WHERE r.lastupdated <> {UPDATE_TAG} WITH r LIMIT {LIMIT_SIZE} DELETE (r) return COUNT(*) as TotalCompleted",
+      "iterative": true,
+      "iterationsize": 100,
+      "__comment__": "Remove GCP Instance-to-Project relationships that are out of date."
+    }
+  ],
+  "name": "cleanup GCP Instances"
+}
