+++ cartography/data/jobs/cleanup/gcp_crm_organization_cleanup.json
@@ -1,0 +1,17 @@
+{
+  "statements": [
+    {
+      "query": "MATCH (n:GCPOrganization) WHERE n.lastupdated <> {UPDATE_TAG} WITH n LIMIT {LIMIT_SIZE} DETACH DELETE (n) return COUNT(*) as TotalCompleted",
+      "iterative": true,
+      "iterationsize": 100,
+      "__comment__": "Remove GCP organizations that are out of date."
+    },
+    {
+      "query": "MATCH (:GCPOrganization)<-[r:PARENT]-(:GCPProject) WHERE r.lastupdated <> {UPDATE_TAG} WITH r LIMIT {LIMIT_SIZE} DELETE (r) return COUNT(*) as TotalCompleted",
+      "iterative": true,
+      "iterationsize": 100,
+      "__comment__": "Remove GCP Organization relationships that are out of date."
+    }
+  ],
+  "name": "cleanup GCP Organizations"
+}
