+++ cartography/data/jobs/cleanup/gcp_crm_project_cleanup.json
@@ -1,0 +1,23 @@
+{
+  "statements": [
+    {
+      "query": "MATCH (n:GCPProject) WHERE n.lastupdated <> {UPDATE_TAG} WITH n LIMIT {LIMIT_SIZE} DETACH DELETE (n) return COUNT(*) as TotalCompleted",
+      "iterative": true,
+      "iterationsize": 100,
+      "__comment__": "Delete GCP Projects that no longer exist and detach them from all previously connected nodes."
+    },
+    {
+      "query": "MATCH (:GCPProject)-[r:PARENT]->(:GCPFolder) WHERE r.lastupdated <> {UPDATE_TAG} WITH r LIMIT {LIMIT_SIZE} DELETE (r) return COUNT(*) as TotalCompleted",
+      "iterative": true,
+      "iterationsize": 100,
+      "__comment__": "Remove GCP Project-to-Folder relationships that are out of date."
+    },
+    {
+      "query": "MATCH (:GCPProject)-[r:PARENT]->(:GCPOrganization) WHERE r.lastupdated <> {UPDATE_TAG} WITH r LIMIT {LIMIT_SIZE} DELETE (r) return COUNT(*) as TotalCompleted",
+      "iterative": true,
+      "iterationsize": 100,
+      "__comment__": "Remove GCP Project-to-Organization relationships that are out of date."
+    }
+  ],
+  "name": "cleanup GCP Projects"
+}
