--- cartography/data/jobs/cleanup/gcp_organization_cleanup.json
@@ -1,9 +1,0 @@
-{
-  "statements": [
-    {
-      "query": "MATCH (n:GCPOrganization) WHERE n.lastupdated <> {UPDATE_TAG} WITH n LIMIT {LIMIT_SIZE} DETACH DELETE (n) return COUNT(*) as TotalCompleted",
-      "iterative": true,
-      "iterationsize": 100
-  }],
-  "name": "cleanup GCP Organizations"
-}
